# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "云盘系统后端 API 文档",
        "title": "CloudDrive API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/share": {
            "delete": {
                "description": "取消指定的分享链接（token或resource_id），仅分享创建者可操作，需登录（Session）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分享"
                ],
                "summary": "取消分享",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分享Token",
                        "name": "token",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "资源ID",
                        "name": "resource_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/share/download/{token}": {
            "get": {
                "description": "通过分享链接下载文件，私有需access_code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "分享"
                ],
                "summary": "分享文件下载",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分享Token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "访问码(私有分享)",
                        "name": "access_code",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/share/private": {
            "get": {
                "description": "查询指定文件的未过期私有分享链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分享"
                ],
                "summary": "查询已有未过期的私有分享",
                "parameters": [
                    {
                        "type": "string",
                        "description": "资源ID",
                        "name": "resource_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.PrivateShareResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个私有分享链接，需访问码访问",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分享"
                ],
                "summary": "创建私有分享链接",
                "parameters": [
                    {
                        "description": "私有分享参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PrivateShareRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.PrivateShareResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/share/public": {
            "get": {
                "description": "查询指定文件的未过期公开分享链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分享"
                ],
                "summary": "查询已有未过期的公开分享",
                "parameters": [
                    {
                        "type": "string",
                        "description": "资源ID",
                        "name": "resource_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.PublicShareResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个公开分享链接，任何人可访问",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分享"
                ],
                "summary": "创建公开分享链接",
                "parameters": [
                    {
                        "description": "公开分享参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PublicShareRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.PublicShareResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/share/{token}": {
            "get": {
                "description": "访问公开或私有分享链接，私有需access_code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分享"
                ],
                "summary": "访问分享链接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分享Token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "访问码(私有分享)",
                        "name": "access_code",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.PublicShareAccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.PrivateShareRequest": {
            "type": "object",
            "required": [
                "expire_hours",
                "resource_id"
            ],
            "properties": {
                "expire_hours": {
                    "type": "integer",
                    "maximum": 168,
                    "minimum": 1
                },
                "resource_id": {
                    "type": "string"
                }
            }
        },
        "handler.PrivateShareResponse": {
            "type": "object",
            "properties": {
                "access_code": {
                    "type": "string"
                },
                "expire_at": {
                    "type": "integer"
                },
                "share_link": {
                    "type": "string"
                }
            }
        },
        "handler.PublicShareAccessResponse": {
            "type": "object",
            "properties": {
                "expire_at": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "integer"
                },
                "resource_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "handler.PublicShareRequest": {
            "type": "object",
            "required": [
                "expire_hours",
                "resource_id"
            ],
            "properties": {
                "expire_hours": {
                    "description": "1~168小时",
                    "type": "integer",
                    "maximum": 168,
                    "minimum": 1
                },
                "resource_id": {
                    "type": "string"
                }
            }
        },
        "handler.PublicShareResponse": {
            "type": "object",
            "properties": {
                "expire_at": {
                    "type": "integer"
                },
                "share_link": {
                    "type": "string"
                }
            }
        },
...
    }
}